support canReturnNullResponseIfMatchNoDocs
